+++ public/services/LogTypeService.ts
@@ -41,7 +41,7 @@ export default class LogTypeService {
             must: {
               query_string: {
                 query:
-                  '(space: Sigma and !(name: others*) and !(name: test*)) or (space: Custom)',
+                  '(space: standard and !(name: others*) and !(name: test*)) or (space: custom)',
               },
             },
           },